+++ jestConfig.js
@@ -1,2 +1,2 @@
 // eslint-disable-next-line import/no-internal-modules
-import "jest-dom/extend-expect";
+import "@testing-library/jest-dom/extend-expect";