+++ packages/components/src/Button/Button.tsx
@@ -1,11 +1,17 @@
 import React from "react";
 import classnames from "classnames";
 import { XOR } from "ts-xor";
+import { Link } from "react-router-dom";
 import styles from "./Button.css";
 import { Typography } from "../Typography";
 import { Icon, IconNames } from "../Icon";
 
 interface ButtonFoundationProps {
+  /**
+   * Used for screen readers. Will override label on screen
+   * reader if present.
+   */
+  readonly ariaLabel?: string;
   readonly ariaControls?: string;
   readonly ariaHaspopup?: boolean;
   readonly ariaExpanded?: boolean;
@@ -15,11 +21,33 @@ interface ButtonFoundationProps {
   readonly icon?: IconNames;
   readonly iconOnRight?: boolean;
   readonly id?: string;
-  readonly label: string;
+  readonly label?: string;
   readonly loading?: boolean;
   readonly size?: "small" | "base" | "large";
-  readonly url?: string;
   onClick?(): void;
+}
+
+interface ButtonIconProps extends ButtonFoundationProps {
+  readonly icon: IconNames;
+  readonly ariaLabel: string;
+}
+
+interface ButtonLabelProps extends ButtonFoundationProps {
+  readonly label: string;
+}
+
+interface ButtonAnchorProps extends ButtonFoundationProps {
+  /**
+   * Used to create an 'href' on an anchor tag.
+   */
+  readonly url?: string;
+}
+
+interface ButtonLinkProps extends ButtonFoundationProps {
+  /**
+   * Used for client side routing. Only use when inside a routed component.
+   */
+  readonly to?: string;
 }
 
 interface BaseActionProps extends ButtonFoundationProps {
@@ -38,30 +66,36 @@ interface CancelActionProps extends ButtonFoundationProps {
 }
 
 export type ButtonProps = XOR<
-  ButtonFoundationProps,
-  XOR<BaseActionProps, XOR<DestructiveActionProps, CancelActionProps>>
->;
+  BaseActionProps,
+  XOR<DestructiveActionProps, CancelActionProps>
+> &
+  XOR<ButtonLinkProps, ButtonAnchorProps> &
+  XOR<ButtonIconProps, ButtonLabelProps>;
 
-export function Button({
-  ariaControls,
-  ariaHaspopup,
-  ariaExpanded,
-  disabled = false,
-  external,
-  fullWidth,
-  icon,
-  iconOnRight,
-  id,
-  label,
-  loading,
-  onClick,
-  size = "base",
-  type = "primary",
-  url,
-  variation = "work",
-}: ButtonProps) {
+export function Button(props: ButtonProps) {
+  const {
+    ariaControls,
+    ariaHaspopup,
+    ariaExpanded,
+    ariaLabel,
+    disabled = false,
+    external,
+    fullWidth,
+    icon,
+    label,
+    iconOnRight,
+    id,
+    loading,
+    onClick,
+    size = "base",
+    type = "primary",
+    url,
+    to,
+    variation = "work",
+  } = props;
+
   const buttonClassNames = classnames(styles.button, styles[size], {
-    [styles.hasIcon]: icon,
+    [styles.hasIconAndLabel]: icon && label,
     [styles.iconOnRight]: iconOnRight,
     [styles[variation]]: variation,
     [styles[type]]: type,
@@ -70,24 +104,46 @@ export function Button({
     [styles.loading]: loading,
   });
 
-  const props = {
+  const tagProps = {
     className: buttonClassNames,
-    disabled: disabled,
-    id: id,
+    disabled,
+    id,
     ...(!disabled && { href: url }),
     ...(!disabled && { onClick: onClick }),
     ...(external && { target: "_blank" }),
-    ...(url === undefined && { type: "button" as "button" }),
+    ...(url === undefined &&
+      to === undefined && { type: "button" as "button" }),
+    "aria-controls": ariaControls,
+    "aria-haspopup": ariaHaspopup,
+    "aria-expanded": ariaExpanded,
+    "aria-label": ariaLabel,
   };
 
+  const buttonInternals = <ButtonInternals {...props} />;
+
+  if (to) {
+    return (
+      <Link {...tagProps} to={to}>
+        {buttonInternals}
+      </Link>
+    );
+  }
+
   const Tag = url ? "a" : "button";
+
+  return <Tag {...tagProps}>{buttonInternals}</Tag>;
+}
+
+function ButtonInternals({
+  label,
+  icon,
+  variation = "work",
+  type = "primary",
+  disabled,
+  size = "base",
+}: ButtonProps) {
   return (
-    <Tag
-      {...props}
-      aria-controls={ariaControls}
-      aria-haspopup={ariaHaspopup}
-      aria-expanded={ariaExpanded}
-    >
+    <>
       {icon && (
         <Icon
           name={icon}
@@ -104,7 +160,7 @@ export function Button({
       >
         {label}
       </Typography>
-    </Tag>
+    </>
   );
 }