+++ packages/components/src/DescriptionList/DescriptionList.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { ReactNode } from "react";
 import styles from "./DescriptionList.css";
 import { Typography } from "../Typography";
 
@@ -7,7 +7,7 @@ interface DescriptionListProps {
    * A tuple where the first item is the string to display as the term
    * and the second value is the string to display as the definition
    */
-  data: [string, string][];
+  data: [string, string | ReactNode][];
 }
 
 export function DescriptionList({ data }: DescriptionListProps) {